add capistrano-mountaintop + github deployments
capistrano mountaintop interacts with the campfire API via capistrano,
everytime you deploy it will be logged in campfire.

also, instead of pushing from your local repo to the server, we now pull
from github. this may have to change when we switch to heroku, but for
the time being it's a good strategy in my opinion.